+++ listeners/commandListener.lua
@@ -1,19 +1,32 @@
 
 include("command-parser/parse-args.lua");
-include("security.lua");
-include("table-ext/table-ext.lua");
 
 
 local textureAliasMap = {
   air = 0,
   dirt = 1,
   grass = 2,
-  stone = 3
-  --- TODO : add more textures!
+  stone = 3,
+  gravel = 4,
+  rock = 5,
+  farmland = 6,
+  mud = 7,
+  snow = 8,
+  sand = 9,
+  desertdirt = 10,
+  desertstone = 11,
+  clay = 12,
+  dungeonwall = 13,
+  dungeonfloor = 14,
+  bonewall = 15,
+  hellstone = 16
 };
 local fillAvailableArgs = table.keys(textureAliasMap);
 local clearAvailableArgs = {"obj","con","veg","all","abs"};
 
+table.sort(fillAvailableArgs, function (a, b)
+  return textureAliasMap[a] < textureAliasMap[b];
+end);
 table.insert(fillAvailableArgs, "#id");
 
 
@@ -21,30 +34,40 @@ local function weHelp(event, args)
   local helpContext = string.lower(args[1] or "");
 
   if helpContext == "select" then
-    print("Showing /we select help");
+    --print("Showing /we select help");
     event.player:sendTextMessage("[#33FF33]/we select");
-    event.player:sendTextMessage("[#FFFF00]"..i18n.t(event.player, "help.select.usage"));
+    for k,line in pairs(string.wrap(i18n.t(event.player, "help.select.usage"), 80)) do
+      event.player:sendTextMessage("[#FFFF00]"..line);
+    end
   elseif helpContext == "cancel" then
-    print("Showing /we cancel help");
+    --print("Showing /we cancel help");
     event.player:sendTextMessage("[#33FF33]/we cancel");
-    event.player:sendTextMessage("[#FFFF00]"..i18n.t(event.player, "help.cancel.usage"));
+    for k,line in pairs(string.wrap(i18n.t(event.player, "help.cancel.usage"), 80)) do
+      event.player:sendTextMessage("[#FFFF00]"..line);
+    end
   elseif helpContext == "clear" then
-    print("Showing /we clear help");
+    --print("Showing /we clear help");
     event.player:sendTextMessage("[#33FF33]/we clear ["..table.concat(clearAvailableArgs, '|').."] [-p]");
-    event.player:sendTextMessage("[#FFFF00]"..i18n.t(event.player, "help.clear.usage"));
+    for k,line in pairs(string.wrap(i18n.t(event.player, "help.clear.usage"), 80)) do
+      event.player:sendTextMessage("[#FFFF00]"..line);
+    end
   elseif helpContext == "fill" then
-    print("Showing /we fill help");
-    event.player:sendTextMessage("[#33FF33]/we fill ["..table.concat(fillAvailableArgs, '|').."] [-p]");
-    event.player:sendTextMessage("[#FFFF00]"..i18n.t(event.player, "help.fill.usage"));
+    --print("Showing /we fill help");
+    event.player:sendTextMessage("[#33FF33]/we fill <texture> [-p]");
+    for k,line in pairs(string.wrap(i18n.t(event.player, "help.fill.usage", table.concat(fillAvailableArgs, ', ')), 80)) do
+      event.player:sendTextMessage("[#FFFF00]"..line);
+    end
   elseif helpContext == "place" then
-    print("Showing /we place help");
-    event.player:sendTextMessage("[#33FF33]/we place blocktype id [north|east|south|west] [sideway|flipped] [-p]");
-    event.player:sendTextMessage("[#FFFF00]"..i18n.t(event.player, "help.place.usage"));
-    event.player:sendTextMessage("[#FFFF00]"..i18n.t(event.player, "help.place.blocktypes", "blocktype", table.concat(getBlockTypes(), ', ')));
+    --print("Showing /we place help");
+    event.player:sendTextMessage("[#33FF33]/we place <blocktype> <id> [north|east|south|west] [sideway|flipped] [-p]");
+    for k,line in pairs(string.wrap(i18n.t(event.player, "help.place.usage", "blocktype", table.concat(getBlockTypes(), ', ')), 80)) do
+      event.player:sendTextMessage("[#FFFF00]"..line);
+    end
   else
-    print("Showing general help");
-    event.player:sendTextMessage("[#33FF33]/we help|select|cancel|clear|fill|place [args]");
-    event.player:sendTextMessage("[#FFFF00]"..i18n.t(event.player, "help.usage", "/we help clear"));
+    event.player:sendTextMessage("[#33FF33]/we <help|select|cancel|clear|fill|place> [args]");
+    for k,line in pairs(string.wrap(i18n.t(event.player, "help.usage", "/we help fill"), 80)) do
+      event.player:sendTextMessage("[#FFFF00]"..line);
+    end
   end
 end
 
@@ -66,7 +89,7 @@ end
 
 local function weSelect(event)
   event.player:enableMarkingSelector(function()
-    print("Area selection start");
+    --print("Area selection start");
     setLabel(event, i18n.t(event.player, "select.start"));
   end);
 end
@@ -77,7 +100,7 @@ local function weCancel(event)
     setLabel(event);
 
     if markingEvent ~= false then
-      print("Area selection cancelled");
+      --print("Area selection cancelled");
       event.player:sendTextMessage("[#FF0000]"..i18n.t(event.player, "select.cancelled"));
     end
   end);
@@ -92,19 +115,19 @@ local function weClear(event, args, flags)
       local coords = getCoordsFromMarkEvent(markingEvent);
 
       if clearObjType == "obj" then
-        print("Clearing area of objects");
+        --print("Clearing area of objects");
         removeObjects(coords);
       elseif clearObjType == "con" then
-        print("Clearing area of construction");
+        --print("Clearing area of construction");
         removeConstr(coords);
       elseif clearObjType == "veg" then
-        print("Clearing area of vegetation");
+        --print("Clearing area of vegetation");
         removeVeg(coords);
       elseif clearObjType == "all" then
-        print("Clearing area of all");
+        --print("Clearing area of all");
         removeAll(coords, false);
       elseif clearObjType == "abs" then
-        print("Clearing area of absolutely everything");
+        --print("Clearing area of absolutely everything");
         removeAll(coords, true);
       else
         return event.player:sendTextMessage("[#FF0000]"..i18n.t(event.player, "cmd.use.args", table.concat(clearAvailableArgs, ", ")));
@@ -132,7 +155,7 @@ local function wePlaceBlock(event, args, flags)
   if blockType and blockId then
 
     if blockId < 21 then
-      print("Block id adjusted from "..blockId.." to 21");
+      --print("Block id adjusted from "..blockId.." to 21");
       blockId = 21;
     end;
 
@@ -140,7 +163,7 @@ local function wePlaceBlock(event, args, flags)
       if markingEvent ~= false then
         local coords = getCoordsFromMarkEvent(markingEvent);
 
-        print("Placing "..blockType.." in area with id "..blockId..(cleanup ~= nil and " with cleanup" or ""));
+        --print("Placing "..blockType.." in area with id "..blockId..(cleanup ~= nil and " with cleanup" or ""));
         fillWithBlock(coords, blockId);
 
         if not flags["p"] then
@@ -167,7 +190,7 @@ local function weFill(event, args, flags)
   if id then
 
     if id < 0 or id > 16 then
-      print("Terrain id adjusted from "..id.." to 0");
+      --print("Terrain id adjusted from "..id.." to 0");
       id = 0;
     end;
 
@@ -175,7 +198,7 @@ local function weFill(event, args, flags)
       if markingEvent ~= false then
         local coords = getCoordsFromMarkEvent(markingEvent);
 
-        print("Filling area with id "..id..(cleanup ~= nil and " with cleanup" or ""));
+        --print("Filling area with id "..id..(cleanup ~= nil and " with cleanup" or ""));
         fillWith(coords, id, cleanup);
 
         if not flags["p"] then
@@ -219,7 +242,6 @@ local function onPlayerCommand(event)
       elseif cmd == "place" then
         if checkPlayerAccess(event.player, "place") then wePlaceBlock(event, table.slice(args, 3), flags); end;
       else
-        print("Unknown command: "..event.command);
         event.player:sendTextMessage("[#FF0000]"..i18n.t(event.player, "cmd.unknown"));
       end
     end